+++ types/http/parseRetryAfter.d.ts
@@ -1,0 +1,1 @@
+export { parseRetryAfter, parseRetryAfter as default } from "../http.js";